check if configmap exist before trying to create it

+++ Test/Altinn.Correspondence.LoadTests/correspondence/run-test-in-k8s.sh
@@ -135,10 +135,12 @@ if [ ! -f "archive.tar" ]; then
 fi
 
 # Create the configmap from the archive
-if ! kubectl create configmap $configmapname --from-file=archive.tar; then
-    echo "Error: Failed to create configmap"
-    rm archive.tar
-    exit 1
+if ! kubectl get configmap $configmapname &>/dev/null; then
+  if ! kubectl create configmap $configmapname --from-file=archive.tar; then
+      echo "Error: Failed to create configmap"
+      rm archive.tar
+      exit 1
+  fi
 fi
 
 # Create the config.yml file from a string
